clean up states
