+++ frontend/scripts/settings.js
@@ -2,7 +2,7 @@ import { volume, setVolume } from './state.js'
 
 const volumeElmmt = document.getElementById('volume');
 
-// volumeElmmt?.value = volume();
+volumeElmmt?.value = volume();
 
 volumeElmmt?.addEventListener("change", (e) => {
     setVolume(e.target.value)
